Move config files into core folder
