+++ alembic/env.py
@@ -4,7 +4,7 @@ from logging.config import fileConfig
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 from alembic import context
-from settings import settings
+from core.settings import settings
 from models import Base
