+++ core/database.py
@@ -21,7 +21,7 @@ def init_db(database_url: str):
     SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 def get_db():
-    from database import SessionLocal  # ensure it's set
+    from core.database import SessionLocal  # ensure it's set
     db = SessionLocal()
     try:
         yield db
