+++ models/__init__.py
@@ -3,4 +3,4 @@ from .user import User
 from .event import Event
 from .user_event import UserEvent
 from .announcement import Announcement
-from database import Base
+from core.database import Base
