+++ models/announcement.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Integer, Text, Text, DateTime
-from database import Base
+from core.database import Base
 
 class Announcement(Base):
     __tablename__ = "announcements"
