+++ tests/settings_test.py
@@ -3,7 +3,7 @@ import os
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 import pytest
-from settings import Settings
+from core.settings import Settings
 
 @pytest.fixture(autouse=True)
 def clear_env_vars():
